docs(compute-samples): added parallel testing support
