+++ compute/cloud-client/src/main/java/compute/CreateFirewallRule.java
@@ -26,11 +26,13 @@ import com.google.cloud.compute.v1.InsertFirewallRequest;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class CreateFirewallRule {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample
     /* project: project ID or project number of the Cloud project you want to use.
        firewallRuleName: name of the rule that is created.
@@ -49,7 +51,7 @@ public class CreateFirewallRule {
   // Creates a simple firewall rule allowing for incoming HTTP and 
   // HTTPS access from the entire Internet.
   public static void createFirewall(String project, String firewallRuleName, String network)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     /* Initialize client that will be used to send requests. This client only needs to be created
        once, and can be reused for multiple requests. After completing all of your requests, call
        the `firewallsClient.close()` method on the client to safely
@@ -78,7 +80,8 @@ public class CreateFirewallRule {
           .setFirewallResource(firewallRule)
           .setProject(project).build();
 
-      firewallsClient.insertAsync(insertFirewallRequest).get();
+      firewallsClient.insertAsync(insertFirewallRequest).get(3, TimeUnit.MINUTES);
+      ;
 
       System.out.println("Firewall rule created successfully -> " + firewallRuleName);
     }
