+++ compute/cloud-client/src/main/java/compute/CreateInstanceTemplate.java
@@ -29,11 +29,13 @@ import com.google.cloud.compute.v1.NetworkInterface;
 import com.google.cloud.compute.v1.Operation;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class CreateInstanceTemplate {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample.
     // projectId: project ID or project number of the Cloud project you use.
     // templateName: name of the new template to create.
@@ -47,7 +49,7 @@ public class CreateInstanceTemplate {
     instance configuration.
    */
   public static void createInstanceTemplate(String projectId, String templateName)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (InstanceTemplatesClient instanceTemplatesClient = InstanceTemplatesClient.create()) {
 
       String machineType = "e2-standard-4";
@@ -85,7 +87,9 @@ public class CreateInstanceTemplate {
               .setProperties(instanceProperties).build()).build();
 
       // Create the Instance Template.
-      Operation response = instanceTemplatesClient.insertAsync(insertInstanceTemplateRequest).get();
+      Operation response = instanceTemplatesClient.insertAsync(insertInstanceTemplateRequest)
+          .get(3, TimeUnit.MINUTES);
+      ;
 
       if (response.hasError()) {
         System.out.println("Instance Template creation failed ! ! " + response);
@@ -97,7 +101,7 @@ public class CreateInstanceTemplate {
   }
 
   public static void createInstanceTemplateWithDiskType(String projectId, String templateName)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (InstanceTemplatesClient instanceTemplatesClient = InstanceTemplatesClient.create();
         GlobalOperationsClient globalOperationsClient = GlobalOperationsClient.create()) {
 
@@ -122,7 +126,9 @@ public class CreateInstanceTemplate {
           .setProject(projectId)
           .setInstanceTemplateResource(instanceTemplate).build();
 
-      Operation response = instanceTemplatesClient.insertAsync(insertInstanceTemplateRequest).get();
+      Operation response = instanceTemplatesClient.insertAsync(insertInstanceTemplateRequest)
+          .get(3, TimeUnit.MINUTES);
+      ;
 
       if (response.hasError()) {
         System.out.println("Instance Template creation failed ! ! " + response);
