+++ compute/cloud-client/src/main/java/compute/DeleteFirewallRule.java
@@ -24,11 +24,13 @@ import com.google.cloud.compute.v1.Operation;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class DeleteFirewallRule {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample
     // project: project ID or project number of the Cloud project you want to use.
     // firewallRuleName: name of the firewall rule you want to delete.
@@ -40,7 +42,7 @@ public class DeleteFirewallRule {
 
   // Deletes a firewall rule from the project.
   public static void deleteFirewallRule(String project, String firewallRuleName)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     /* Initialize client that will be used to send requests. This client only needs to be created
        once, and can be reused for multiple requests. After completing all of your requests, call
        the `firewallsClient.close()` method on the client to safely
@@ -49,7 +51,8 @@ public class DeleteFirewallRule {
 
       OperationFuture<Operation, Operation> operation = firewallsClient.deleteAsync(project,
           firewallRuleName);
-      operation.get();
+      operation.get(3, TimeUnit.MINUTES);
+      ;
 
       System.out.println("Deleted firewall rule -> " + firewallRuleName);
     }
