+++ compute/cloud-client/src/main/java/compute/PatchFirewallRule.java
@@ -26,11 +26,13 @@ import com.google.cloud.compute.v1.PatchFirewallRequest;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class PatchFirewallRule {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample
     // project: project ID or project number of the Cloud project you want to use.
     // firewallRuleName: name of the rule you want to modify.
@@ -44,7 +46,7 @@ public class PatchFirewallRule {
 
   // Modifies the priority of a given firewall rule.
   public static void patchFirewallPriority(String project, String firewallRuleName, int priority)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     /* Initialize client that will be used to send requests. This client only needs to be created
        once, and can be reused for multiple requests. After completing all of your requests, call
        the `firewallsClient.close()` method on the client to safely
@@ -63,7 +65,8 @@ public class PatchFirewallRule {
 
       OperationFuture<Operation, Operation> operation = firewallsClient.patchAsync(
           patchFirewallRequest);
-      operation.get();
+      operation.get(3, TimeUnit.MINUTES);
+      ;
       System.out.println("Firewall Patch applied successfully ! ");
     }
   }
